fixing a couple bugs

+++ src/test/Problem14.scala
@@ -1,16 +1,16 @@
 package test
 
 object Problem14 extends App {
-  //476 length
-  //910107 number
-  var max = 0;
-  var maxNum = 0;
-  for (n <- 1 to 1000000) {
+  //525 length
+  //837799 number
+  var max = 0L;
+  var maxNum = 0L;
+  for (n <- 1L to 1000000L) {
     var chain = 1;
     var m = n;
     while (m > 1) {
       if (m % 2 == 0) {
-        m = m / 2;
+        m = (m / 2).toLong;
         chain = chain + 1;
       } else {
         m = 3 * m + 1;
